@@ -1,11 +1,3 @@
-namespace FineCollectionService.Models
-{
-    public class VehicleInfo
-    {
-        public string VehicleId { get; set; }
-        public string Brand { get; set; }
-        public string Model { get; set; }
-        public string OwnerName { get; set; }
-        public string OwnerEmail { get; set; }
-    }
-}
+namespace FineCollectionService.Models;
+
+public record struct VehicleInfo(string VehicleId, string Brand, string Model, string OwnerName, string OwnerEmail);
