@@ -1,11 +1,7 @@
-using System;
-
-namespace Simulation.Events
+namespace Simulation.Events;
+public class VehicleRegistered
 {
-    public class VehicleRegistered
-    {
-        public int Lane { get; set; }
-        public string LicenseNumber { get; set; }
-        public DateTime Timestamp { get; set; }
-    }
+    public int Lane { get; set; }
+    public string LicenseNumber { get; set; }
+    public DateTime Timestamp { get; set; }
 }
