@@ -16,8 +16,7 @@ namespace FineCollectionService.Proxies
 
         public async Task<VehicleInfo> GetVehicleInfo(string licenseNumber)
         {
-            return await _httpClient.GetFromJsonAsync<VehicleInfo>(
-                $"http://localhost:6002/vehicleinfo/{licenseNumber}");
+            return await _httpClient.GetFromJsonAsync<VehicleInfo>($"http://localhost:6002/vehicleinfo/{licenseNumber}");
         }       
     }
 }
