@@ -1,10 +1,7 @@
-using Simulation.Events;
+namespace Simulation.Proxies;
 
-namespace Simulation.Proxies
+public interface ITrafficControlService
 {
-    public interface ITrafficControlService
-    {
-        public void SendVehicleEntry(VehicleRegistered vehicleRegistered);
-        public void SendVehicleExit(VehicleRegistered vehicleRegistered);
-    }
+    public void SendVehicleEntry(VehicleRegistered vehicleRegistered);
+    public void SendVehicleExit(VehicleRegistered vehicleRegistered);
 }
