@@ -1,110 +1,98 @@
-﻿using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.Logging;
-using TrafficControlService.Events;
-using TrafficControlService.DomainServices;
-using TrafficControlService.Models;
-using System.Net.Http;
-using System.Net.Http.Json;
-using TrafficControlService.Repositories;
+﻿namespace TrafficControlService.Controllers;
 
-namespace TrafficControlService.Controllers
+[ApiController]
+[Route("")]
+public class TrafficController : ControllerBase
 {
-    [ApiController]
-    [Route("")]
-    public class TrafficController : ControllerBase
+    private readonly HttpClient _httpClient;
+    private readonly IVehicleStateRepository _vehicleStateRepository;
+    private readonly ILogger<TrafficController> _logger;
+    private readonly ISpeedingViolationCalculator _speedingViolationCalculator;
+    private readonly string _roadId;
+
+    public TrafficController(
+        ILogger<TrafficController> logger,
+        HttpClient httpClient,
+        IVehicleStateRepository vehicleStateRepository,
+        ISpeedingViolationCalculator speedingViolationCalculator)
     {
-        private readonly HttpClient _httpClient;
-        private readonly IVehicleStateRepository _vehicleStateRepository;
-        private readonly ILogger<TrafficController> _logger;
-        private readonly ISpeedingViolationCalculator _speedingViolationCalculator;
-        private readonly string _roadId;
+        _logger = logger;
+        _httpClient = httpClient;
+        _vehicleStateRepository = vehicleStateRepository;
+        _speedingViolationCalculator = speedingViolationCalculator;
+        _roadId = speedingViolationCalculator.GetRoadId();
+    }
 
-        public TrafficController(
-            ILogger<TrafficController> logger,
-            HttpClient httpClient,
-            IVehicleStateRepository vehicleStateRepository,
-            ISpeedingViolationCalculator speedingViolationCalculator)
+    [HttpPost("entrycam")]
+    public async Task<ActionResult> VehicleEntry(VehicleRegistered msg)
+    {
+        try
         {
-            _logger = logger;
-            _httpClient = httpClient;
-            _vehicleStateRepository = vehicleStateRepository;
-            _speedingViolationCalculator = speedingViolationCalculator;
-            _roadId = speedingViolationCalculator.GetRoadId();
-        }
+            // log entry
+            _logger.LogInformation($"ENTRY detected in lane {msg.Lane} at {msg.Timestamp.ToString("hh:mm:ss")} " +
+                $"of vehicle with license-number {msg.LicenseNumber}.");
 
-        [HttpPost("entrycam")]
-        public async Task<ActionResult> VehicleEntry(VehicleRegistered msg)
-        {
-            try
+            // store vehicle state
+            var vehicleState = new VehicleState
             {
-                // log entry
-                _logger.LogInformation($"ENTRY detected in lane {msg.Lane} at {msg.Timestamp.ToString("hh:mm:ss")} " +
-                    $"of vehicle with license-number {msg.LicenseNumber}.");
+                LicenseNumber = msg.LicenseNumber,
+                EntryTimestamp = msg.Timestamp
+            };
+            await _vehicleStateRepository.SaveVehicleStateAsync(vehicleState);
 
-                // store vehicle state
-                var vehicleState = new VehicleState
+            return Ok();
+        }
+        catch
+        {
+            return StatusCode(500);
+        }
+    }
+
+    [HttpPost("exitcam")]
+    public async Task<ActionResult> VehicleExit(VehicleRegistered msg)
+    {
+        try
+        {
+            // get vehicle state
+            var state = await _vehicleStateRepository.GetVehicleStateAsync(msg.LicenseNumber);
+            if (state == default(VehicleState))
+            {
+                return NotFound();
+            }
+
+            // log exit
+            _logger.LogInformation($"EXIT detected in lane {msg.Lane} at {msg.Timestamp.ToString("hh:mm:ss")} " +
+                $"of vehicle with license-number {msg.LicenseNumber}.");
+
+            // update state
+            var exitState = state.Value with { ExitTimestamp = msg.Timestamp };
+            await _vehicleStateRepository.SaveVehicleStateAsync(exitState);
+
+            // handle possible speeding violation
+            int violation = _speedingViolationCalculator.DetermineSpeedingViolationInKmh(exitState.EntryTimestamp, exitState.ExitTimestamp.Value);
+            if (violation > 0)
+            {
+                _logger.LogInformation($"Speeding violation detected ({violation} KMh) of vehicle" +
+                    $"with license-number {state.Value.LicenseNumber}.");
+
+                var speedingViolation = new SpeedingViolation
                 {
-                    LicenseNumber = msg.LicenseNumber,
-                    EntryTimestamp = msg.Timestamp
+                    VehicleId = msg.LicenseNumber,
+                    RoadId = _roadId,
+                    ViolationInKmh = violation,
+                    Timestamp = msg.Timestamp
                 };
-                await _vehicleStateRepository.SaveVehicleStateAsync(vehicleState);
 
-                return Ok();
+                // publish speedingviolation
+                var message = JsonContent.Create<SpeedingViolation>(speedingViolation);
+                await _httpClient.PostAsync("http://localhost:6001/collectfine", message);
             }
-            catch
-            {
-                return StatusCode(500);
-            }
+
+            return Ok();
         }
-
-        [HttpPost("exitcam")]
-        public async Task<ActionResult> VehicleExit(VehicleRegistered msg)
+        catch
         {
-            try
-            {
-                // get vehicle state
-                var vehicleState = await _vehicleStateRepository.GetVehicleStateAsync(msg.LicenseNumber);
-                if (vehicleState == null)
-                {
-                    return NotFound();
-                }
-
-                // log exit
-                _logger.LogInformation($"EXIT detected in lane {msg.Lane} at {msg.Timestamp.ToString("hh:mm:ss")} " +
-                    $"of vehicle with license-number {msg.LicenseNumber}.");
-
-                // update state
-                vehicleState.ExitTimestamp = msg.Timestamp;
-                await _vehicleStateRepository.SaveVehicleStateAsync(vehicleState);
-
-                // handle possible speeding violation
-                int violation = _speedingViolationCalculator.DetermineSpeedingViolationInKmh(
-                    vehicleState.EntryTimestamp, vehicleState.ExitTimestamp);
-                if (violation > 0)
-                {
-                    _logger.LogInformation($"Speeding violation detected ({violation} KMh) of vehicle" +
-                        $"with license-number {vehicleState.LicenseNumber}.");
-
-                    var speedingViolation = new SpeedingViolation
-                    {
-                        VehicleId = msg.LicenseNumber,
-                        RoadId = _roadId,
-                        ViolationInKmh = violation,
-                        Timestamp = msg.Timestamp
-                    };
-
-                    // publish speedingviolation
-                    var message = JsonContent.Create<SpeedingViolation>(speedingViolation);
-                    await _httpClient.PostAsync("http://localhost:6001/collectfine", message);
-                }
-
-                return Ok();
-            }
-            catch
-            {
-                return StatusCode(500);
-            }
+            return StatusCode(500);
         }
     }
 }
