@@ -1,11 +1,3 @@
-using System;
+namespace TrafficControlService.Events;
 
-namespace TrafficControlService.Events
-{
-    public class VehicleRegistered
-    {
-        public int Lane { get; set; }
-        public string LicenseNumber { get; set; }
-        public DateTime Timestamp { get; set; }
-    }
-}
+public record struct VehicleRegistered(int Lane, string LicenseNumber, DateTime Timestamp);
