@@ -1,11 +1,15 @@
-using System;
+namespace TrafficControlService.Models;
 
-namespace TrafficControlService.Models
+public record struct VehicleState
 {
-    public class VehicleState
+    public string LicenseNumber { get; init; }
+    public DateTime EntryTimestamp { get; init; }
+    public DateTime? ExitTimestamp { get; init; }
+
+    public VehicleState(string licenseNumber, DateTime entryTimestamp, DateTime? exitTimestamp = null)
     {
-        public string LicenseNumber { get; set; }
-        public DateTime EntryTimestamp { get; set; }
-        public DateTime ExitTimestamp { get; set; }
+        this.LicenseNumber = licenseNumber;
+        this.EntryTimestamp = entryTimestamp;
+        this.ExitTimestamp = exitTimestamp;
     }
 }
