@@ -1,11 +1,7 @@
-using System.Threading.Tasks;
-using TrafficControlService.Models;
+namespace TrafficControlService.Repositories;
 
-namespace TrafficControlService.Repositories
+public interface IVehicleStateRepository
 {
-    public interface IVehicleStateRepository
-    {
-         Task SaveVehicleStateAsync(VehicleState vehicleState);
-         Task<VehicleState> GetVehicleStateAsync(string licenseNumber);
-    }
+        Task SaveVehicleStateAsync(VehicleState vehicleState);
+        Task<VehicleState?> GetVehicleStateAsync(string licenseNumber);
 }
