@@ -1,28 +1,20 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.Hosting;
-using Microsoft.Extensions.Logging;
+// create web-app
+var builder = WebApplication.CreateBuilder(args);
 
-namespace VehicleRegistrationService
+builder.Services.AddScoped<IVehicleInfoRepository, InMemoryVehicleInfoRepository>();
+
+builder.Services.AddControllers();
+
+var app = builder.Build();
+
+// configure web-app
+if (app.Environment.IsDevelopment())
 {
-    public class Program
-    {
-        public static void Main(string[] args)
-        {
-            CreateHostBuilder(args).Build().Run();
-        }
-
-        public static IHostBuilder CreateHostBuilder(string[] args) =>
-            Host.CreateDefaultBuilder(args)
-                .ConfigureWebHostDefaults(webBuilder =>
-                {
-                    webBuilder
-                        .UseUrls("http://localhost:6002")
-                        .UseStartup<Startup>();
-                });
-    }
+    app.UseDeveloperExceptionPage();
 }
+
+// configure routing
+app.MapControllers();
+
+// let's go!
+app.Run();
